program-data-api: Bucket takes precedence over diff for current data
Otherwise the final argument to `triggerFieldUpdate' will always be `true',
meaning that data didn't change (when in fact it did).  Ideally, it should
never be the case that there's a value in the diff that isn't in the bucket,
but it could happen.

This potential unearths old bugs from 2013/2014 (internal id: FS#14712), but
hopefully those situations no longer apply after various changes to liza
since.

* src/program-data-api.xsl (compiler:compile-api-init)[lv:question/lv:data]:
  Give bucket precedence over diff for `cdata'.

DEV-4077